requirements.txt
# 麦田智囊 - 乡村教育Agent系统 v1.0.0

# 核心依赖
streamlit>=1.28.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0

# 嵌入模型
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0

# OCR
easyocr>=1.7.0

# 工具库
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# 数据处理
pillow>=10.0.0
numpy>=1.24.0

# 开发依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
